backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.0
python-docx==1.1.2
pdfplumber==0.11.3
openai==1.45.0
python-dotenv==1.0.1
reportlab==4.2.4
